Refactored SliceRule to be independent of concrete ArchCondition
